Move tzlist loading code to new module
